feat!: Store 4xx responses as-is
Previously the proxy server would replace the response body of 4xx
responses with its own error structure. This breaks GraphQL APQ[1] which
relies on a specific error code to be present in the response body from
a 404 response.

This change removes the special treatment of all 4xx errors and caches
4xx responses as-is without modification.

[1] https://www.apollographql.com/docs/apollo-server/performance/apq